Finalização dos testes unitários

+++ CopaDeFilmesApi/CopaDeFilmes.Dominio/Filme.cs
@@ -5,6 +5,6 @@
         public string Id { get; set; }
         public string Titulo { get; set; }
         public int Ano { get; set; }
-        public float Nota { get; set; }
+        public double Nota { get; set; }
     }
 }
